Build the app in a factory and split the routes into blueprints
main.py created the app, opened the database and registered all sixteen
routes as a side effect of being imported. create_app() now builds a
configured app on demand, and the routes live in auth, catalogue, basket
and reviews blueprints. Every URL and HTTP method is unchanged, and
test_routes.py pins them so a moved URL fails the unit tests.

The database URL moves into config. Tests get a fresh app per test backed
by an in-memory database instead of sharing mydb.db.
