+++ Cloth Shop Website/Web/blueprints/__init__.py
@@ -1,0 +1,7 @@
+from flask import current_app
+
+from database import SessionFactory
+
+
+def db_Session() -> SessionFactory:
+    return current_app.extensions['db_Session']
